+++ dbsync/lib/dbsync.dart
@@ -14,3 +14,4 @@ export 'src/local_entity_schema.dart';
 export 'src/local_changes_schema.dart';
 export 'src/upload_synchronizer.dart';
 export 'src/download_synchronizer.dart';
+export 'src/grpc_sync.dart';