+++ dbsync/lib/src/sync_entity_repository.dart
@@ -1,30 +1,41 @@
+import 'package:meta/meta.dart';
 import 'package:sqflite_common/sqlite_api.dart';
 import 'package:dbsync/dbsync.dart';
 
 abstract class SyncEntityRepository<TEntity> {
   const SyncEntityRepository({
     required this.syncHandler,
-    // required this.syncLocalRepository,
     required this.database,
     required this.isOffline,
   });
 
   final SyncTypeHandler<TEntity> syncHandler;
-  // final SyncLocalRepository syncLocalRepository;
   final Database database;
   final bool isOffline;
 
   Future<TEntity> get(String id) async {
-    if (!isOffline) return await syncHandler.getRemote(id);
+    if (!isOffline) {
+      final remote = await getRemote(id);
+      if (remote != null) {
+        return remote;
+      }
+    }
     return await syncHandler.getLocal(database, id);
   }
 
+  @protected
+  Future<TEntity?> getRemote(String id) async {
+    final e = await syncHandler.getRemote(id);
+    return e;
+  }
+
   Future<TEntity> create(TEntity entity) async {
-    final created = isOffline ? entity : await syncHandler.createRemote(entity);
+    final remoteCreated = isOffline ? null : await createRemote(entity);
+    final created = remoteCreated ?? entity;
 
     await database.transaction((txn) async {
       await syncHandler.upsertLocal(txn, entity);
-      if (isOffline) {
+      if (remoteCreated == null) {
         final localChange = LocalChange.create(
           protoBytes: syncHandler.marshal(entity),
           entityType: TEntity,
@@ -38,12 +49,19 @@ abstract class SyncEntityRepository<TEntity> {
     return created;
   }
 
+  @protected
+  Future<TEntity?> createRemote(TEntity entity) async {
+    final created = await syncHandler.createRemote(entity);
+    return created;
+  }
+
   Future<TEntity> update(TEntity entity) async {
-    final updated = isOffline ? entity : await syncHandler.updateRemote(entity);
+    TEntity? remoteUpdated = isOffline ? null : await updateRemote(entity);
+    final updated = remoteUpdated ?? entity;
 
     await database.transaction((txn) async {
       await syncHandler.upsertLocal(txn, updated);
-      if (isOffline) {
+      if (remoteUpdated == null) {
         final localChange = LocalChange.update(
           entityType: TEntity,
           protoBytes: syncHandler.marshal(entity),
@@ -57,14 +75,23 @@ abstract class SyncEntityRepository<TEntity> {
     return updated;
   }
 
+  /// Tries to update remotely.
+  /// Returns:
+  /// - The updated entity if succeed
+  /// - null if unavailable
+  /// throws if any other exception
+  @protected
+  Future<TEntity?> updateRemote(TEntity entity) async {
+    final updated = await syncHandler.updateRemote(entity);
+    return updated;
+  }
+
   Future<void> delete(String id, String rev) async {
-    if (!isOffline) {
-      await syncHandler.deleteRemote(id, rev);
-    }
+    final synced = isOffline ? false : await deleteRemote(id, rev);
 
     await database.transaction((txn) async {
       await syncHandler.deleteLocal(txn, id);
-      if (isOffline) {
+      if (!synced) {
         final localChange = LocalChange.delete(
           entityType: TEntity,
           entityId: id,
@@ -73,5 +100,15 @@ abstract class SyncEntityRepository<TEntity> {
         await SyncLocalRepository.insertChange(txn, localChange);
       }
     });
+  }
+
+  // Tries to delete the entity remotely
+  // Returns:
+  // - true if succeeded
+  // - false if unavailable
+  // throws if any other exception
+  Future<bool> deleteRemote(String id, String rev) async {
+    await syncHandler.deleteRemote(id, rev);
+    return true;
   }
 }